seen and watchlist page functional, trying to avoid duplicated code so search does not work atm

+++ server/views.py
@@ -60,9 +60,11 @@ def login(request):
     ref.update({'email': email})
     ref.update({'displayName': displayName})
 
-    watchlist = wl if (wl:=ref.child('watchlist').get()) else []
+    watchlist = wl if (wl:=ref.child('watchlist').get()) else {}
+    seenlist = sl if (sl:=ref.child('seenlist').get()) else {}
     print("watchlist:\n", watchlist)
-    return Response({'uid': uid, 'watchlist': watchlist}, status=200)
+    print("seenlist:\n", seenlist)
+    return Response({'uid': uid, 'watchlist': watchlist, 'seenlist': seenlist}, status=200)
 
 
 @api_view(["GET"])
@@ -123,6 +125,45 @@ def remove_watchlist_item(request, movie_id):
     decoded_token = auth.verify_id_token(id_token, check_revoked=True)
 
     ref = db.reference('Users').child(decoded_token['uid']).child('watchlist')
+    ref.child(movie_id).delete()
+
+    return Response({'result': "DELETE"}, status=200)
+
+@api_view(["POST"])
+def add_seenlist_item(request):
+    id_token = request.data['idToken']
+    if not id_token:
+        return Response({'error': "No authentication token."}, status=404)
+
+    decoded_token = auth.verify_id_token(id_token, check_revoked=True)
+
+    if not request.data.get('movie'):
+        return Response({'error': "Movie was not found."}, status=404)
+
+    ref = db.reference('Users').child(decoded_token['uid']).child('seenlist')
+    print(request.data['movie'])
+    ref.update({request.data['movie']['imdb_id']: request.data['movie']})
+
+
+    return Response({'result': "POST"}, status=200)
+
+
+@api_view(["DELETE"])
+def remove_seenlist_item(request, movie_id):
+
+    # id_token = request.headers.get("Authentication")
+    id_token = request.META.get('HTTP_AUTHORIZATION')
+    # print(id_token)
+
+    if not id_token:
+        return Response({'error': "No authentication token."}, status=401)
+
+    if not movie_id:
+        return Response({'error': "You don't have this movie in your watchlist."}, status=404)
+
+    decoded_token = auth.verify_id_token(id_token, check_revoked=True)
+
+    ref = db.reference('Users').child(decoded_token['uid']).child('seenlist')
     ref.child(movie_id).delete()
 
     return Response({'result': "DELETE"}, status=200)
